Bug fixes, updating mapping code
Added uber filter to just call everything at once.

Fixed the mapping definition so that it now plays nicely with the sunpy
metadata.

Moved example notebooks into the notebooks directory.

Map_Example.ipynb is now complete. Need to add the conversion to solar
coordinates in as a module and refactor the
Convert_to_Solar_Example.ipynb