Updates to the planning code
make_mosaic now has an optional extra_roll keyword (which needs to have some astropy angle as its units) that will rotate the mosaic *clockwise* around the center of the Sun. This allows us to specify the PA angle in cases where there may be a CHU blockage and/or the freedom to rotate the mosaic into a diamond or some other shape if we so choose.